requirements.txt
beautifulsoup4 @ file:///tmp/build/80754af9/beautifulsoup4_1631874778482/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1633534655931/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1635944169458/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1634667019719/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mysql-connector-python==8.0.28
pandas==1.3.4
regex @ file:///tmp/build/80754af9/regex_1628063347754/work
requests==2.24.0
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work